feat(u-15b): Ops/Gateway 스키마 v2 전환 + 계약 스냅샷 테스트
Cards API에 Pydantic v2 스키마 적용 및 응답 구조 안정성 보장.

## 구현
- apps/ops/schemas.py (195줄): 7개 Pydantic 모델
  - CardsReasonTrendsResponse: 메인 응답 래퍼
  - CardsDeltaResponse: Delta 프로토콜 래퍼
  - 중첩 모델: ReasonTrendItem, BucketEntry, GroupSummary, CardsSummary, CardsMeta
  - ConfigDict (v2) / Config (v1) 이중 지원

- apps/ops/api/cards.py: response_model 선언, Vary 헤더 추가
  - response_model=Union[CardsReasonTrendsResponse, CardsDeltaResponse]
  - Vary: Accept-Encoding, X-Scopes

## 버그 수정
- apps/ops/api/cards_data.py:107: count 타입 일관성 (float → int)
  - 집계 시 count는 int, score만 float

## 테스트
- tests/contracts/test_router_response_contract_v2.py (6/6 통과):
  1. 응답 구조 검증 (groups, buckets, top_reasons, summary)
  2. ETag/Cache-Control/Vary 헤더 존재
  3. 304 Not Modified 동작
  4. 필드 타입 검증 (int/float/str/dict/list)
  5. JSON 직렬화 검증 (NaN/Inf 없음, 단어 경계 정규식)
  6. Health check 구조

## 결과
- 30/30 테스트 통과 (u-15a: 24 + u-15b: 6)
- API 하위 호환성 100% 보장
- 응답 구조 변경 없음

+++ apps/ops/api/cards_data.py
@@ -104,11 +104,11 @@ def compute_reason_trends(
         bucket_groups = defaultdict(lambda: {"score": 0.0, "count": 0})
         for label, count in reasons.items():
             lbl = str(label)
-            cnt = float(count or 0)
+            cnt = int(count or 0)  # Keep as int for count fields
             grp = _label_group(lbl)
             l_w = label_weights.get(lbl, 1.0)
             g_w = group_weights.get(grp, 1.0)
-            score = cnt * l_w * g_w
+            score = float(cnt) * l_w * g_w  # Use float only for score calculation
             bucket_groups[grp]["score"] += score
             bucket_groups[grp]["count"] += cnt
             groups[grp]["score"] += score
